refactor(shared): 外移 shared_runtime_coordinator.rs 内联测试至 coordinator_tests.rs
- 4864→2670 行，纯搬运零行为变更；#[path] 平铺保私有字段访问
- baseline 重生成：coordinator reduced -2194；新文件登记豁免
验证：cargo test --lib shared_runtime_coordinator 46 passed；cargo check 通过；gate 绿
